Change header title and skip enqueueing state for secure flow
Make start of engagement implicit by skipping enqueueing state and changing title only when engagement is created for secure messaging flow.

MOB-1920
